+++ src/main/java/org/stellar/sdk/requests/sorobanrpc/SimulateTransactionRequest.java
@@ -1,5 +1,6 @@
 package org.stellar.sdk.requests.sorobanrpc;
 
+import java.math.BigInteger;
 import lombok.AllArgsConstructor;
 import lombok.Value;
 
@@ -13,4 +14,12 @@ import lombok.Value;
 @Value
 public class SimulateTransactionRequest {
   String transaction;
+
+  ResourceConfig resourceConfig;
+
+  @Value
+  @AllArgsConstructor
+  public static class ResourceConfig {
+    BigInteger instructionLeeway;
+  }
 }